+++ src/console/controllers/ElasticsearchController.php
@@ -10,18 +10,13 @@
 
 namespace lhs\elasticsearch\console\controllers;
 
-use GuzzleHttp\Client;
-use GuzzleHttp\Exception\ServerException;
-use GuzzleHttp\Pool;
-use GuzzleHttp\Psr7\Request;
-use lhs\elasticsearch\Elasticsearch;
+use Craft;
 use lhs\elasticsearch\Elasticsearch as ElasticsearchPlugin;
 use lhs\elasticsearch\exceptions\IndexElementException;
-use Psr\Http\Message\ResponseInterface;
 use yii\console\Controller;
 use yii\console\ExitCode;
+use yii\helpers\ArrayHelper;
 use yii\helpers\Console;
-use function GuzzleHttp\Psr7\build_query;
 
 /**
  * Manage Craft Elasticsearch indexes from the command line
@@ -29,43 +24,26 @@ use function GuzzleHttp\Psr7\build_query;
 class ElasticsearchController extends Controller
 {
     /**
-     * @var string The public domain name of the site. Will be used to replace the domain name part of the `siteBaseUrl` argument.
-     * This may be useful in advanced server setups (eg. when the server running this CLI script cannot resolve the public domain name).
-     */
-    public $publicDomainName;
-
-    public function options($actionID)
-    {
-        return array_merge(parent::options($actionID), ['publicDomainName']);
-    }
-
-/**
      * Reindex Craft entries into the Elasticsearch instance
      *
-     * @param string $siteBaseUrl The base URL to access the site to index.
-     *                            Should include the protocol.
-     *                            If not http:// will be prepended.
-     *
      * @return int A shell exit code. 0 indicated success, anything else indicates an error
- * @throws IndexElementException If an error occurs while reindexing the entries
+     * @throws IndexElementException If an error occurs while reindexing the entries
+     * @throws \GuzzleHttp\Exception\GuzzleException
+     * @throws \yii\base\InvalidRouteException
+     * @throws \yii\console\Exception
+     * @throws \yii\elasticsearch\Exception
      */
-    public function actionReindexAll($siteBaseUrl): int
+    public function actionReindexAll(): int
     {
         $this->stdout(PHP_EOL);
-        $this->stdout('Craft Elasticsearch plugin | Reindex all entries', Console::FG_GREEN);
+        $this->stdout('Craft Elasticsearch plugin | Reindex all elements', Console::FG_GREEN);
         $this->stdout(PHP_EOL);
 
-        // Ensure that `siteBaseUrl` includes the protocol
-        if (!preg_match('|^https?://.+|', $siteBaseUrl)) {
-            $siteBaseUrl = 'http://' . $siteBaseUrl;
-        }
-        $httpClient = new Client();
+        // Get elements
+        $elements = $this->getElementsToReindex();
 
-        // Get entries
-        $entries = $this->getEntriesToReindex($httpClient, $siteBaseUrl);
-
-        // Reindex entries
-        $exitCode = $this->reindexEntries($httpClient, $siteBaseUrl, $entries);
+        // Reindex elements
+        $exitCode = $this->reindexElements($elements);
 
         // Print summary message
         $this->stdout(PHP_EOL);
@@ -90,127 +68,63 @@ class ElasticsearchController extends Controller
     }
 
     /**
-     * Get an array of associative arrays representing the entries to reindex
-     * (an associative array representing an entry has the following keys: id and url)
-     *
-     * @param Client $httpClient  The instance of Guzzle HTTP Client to use to make the HTTP request
-     * @param string $siteBaseUrl The base URL to access the site to index.
-     *                            Should include the protocol.
-     *                            If not http:// will be prepended.
+     * Get an array of associative arrays representing the elements to reindex
      *
      * @return array An associative arrays representing the entries to reindex
-     * @throws IndexElementException If an error occurs while fetching the list of entries to reindex
      */
-    protected function getEntriesToReindex(Client $httpClient, string $siteBaseUrl): array
+    protected function getElementsToReindex(): array
     {
-        $response = $httpClient->post("{$siteBaseUrl}/elasticsearch/get-all-entries", [
-            'headers' => ['Accept' => 'application/json'],
-        ]);
-
-        $statusCode = $response->getStatusCode();
-        if ($statusCode !== 200) {
-            $reason = $response->getReasonPhrase();
-            throw new IndexElementException("Unexpected response from endpoint: {$statusCode} {$reason}");
+        $elasticsearch = ElasticsearchPlugin::getInstance();
+        $elements = $elasticsearch->service->getEnabledEntries();
+        if ($elasticsearch->isCommerceEnabled()) {
+            $elements = ArrayHelper::merge($elements, $elasticsearch->service->getEnabledProducts());
         }
-
-        $decodedResponse = json_decode($response->getBody());
-        if (json_last_error() !== JSON_ERROR_NONE) {
-            throw new IndexElementException('Invalid JSON received from endpoint.');
-        }
-
-        if (!property_exists($decodedResponse, 'entries')) {
-            throw new IndexElementException('Unexpected JSON content received from endpoint.');
-        }
-
-        return $decodedResponse->entries;
+        return $elements;
     }
 
     /**
-     * Reindex the given $entries. The reindexing process takes place in a web
-     * context as running it in a console context messes up some variables and
-     * URLs.
+     * Reindex the given $entries and show a progress bar.
      *
-     * @param Client $httpClient
-     * @param string $siteBaseUrl The base URL to access the site to index.
-     *                            Should include the protocol.
-     *                            If not http:// will be prepended.
-     * @param array  $entries
+     * @param array $elements
      *
      * @return int A shell exit code. 0 indicated success, anything else indicates an error
+     * @throws IndexElementException
+     * @throws \GuzzleHttp\Exception\GuzzleException
+     * @throws \yii\base\InvalidRouteException
+     * @throws \yii\console\Exception
+     * @throws \yii\elasticsearch\Exception
      */
-    protected function reindexEntries(Client $httpClient, string $siteBaseUrl, array $entries): int
+    protected function reindexElements(array $elements): int
     {
-        $reindexEntryEndpoint = "{$siteBaseUrl}/elasticsearch/reindex-entry";
-
-        $reindexEntryRequests = array_map(function ($entry) use ($reindexEntryEndpoint) {
-            return new Request(
-                'GET',
-                $reindexEntryEndpoint . '?' . build_query(['entryId' => $entry->entryId, 'siteId' => $entry->siteId]),
-                ['Accept' => 'application/json']
-            );
-        }, $entries);
-
-        $entryCount = count($entries);
-        $processedEntryCount = 0;
+        $elementCount = count($elements);
+        $processedElementCount = 0;
         $errorCount = 0;
-        Console::startProgress(0, $entryCount);
+        Console::startProgress(0, $elementCount);
 
-        $reindexEntryRequestPool = new Pool($httpClient, $reindexEntryRequests, [
-            'concurrency' => 5,
-            'fulfilled'   => function () use (&$processedEntryCount, $entryCount) {
-                Console::updateProgress(++$processedEntryCount, $entryCount);
-            },
-            'rejected'    => function (ServerException $reason, $index) use (&$processedEntryCount, $entryCount, &$errorCount) {
+        foreach ($elements as $index => $elementParams) {
+            $element = $elementParams['type'] === 'craft\\commerce\\elements\\Product' ? craft\commerce\Plugin::getInstance()->getProducts()->getProductById($elementParams['elementId'], $elementParams['siteId']) : Craft::$app->getEntries()->getEntryById($elementParams['elementId'], $elementParams['siteId']);
+
+            if ($element === null) {
+                throw new IndexElementException(Craft::t(
+                    ElasticsearchPlugin::TRANSLATION_CATEGORY,
+                    'No such element (element #{elementId} / site #{siteId}',
+                    ['elementId' => $elementParams->elementId, 'siteId' => $elementParams->siteId]
+                ));
+            }
+
+            $result = ElasticsearchPlugin::getInstance()->service->indexElement($element);
+
+            if ($result === null) {
+                Console::updateProgress(++$processedElementCount, $elementCount);
+            } else {
                 $errorCount++;
-                Console::updateProgress(++$processedEntryCount, $entryCount);
-                $this->printError($reason->getResponse(), $index);
-            },
-        ]);
+                Console::updateProgress(++$processedElementCount, $elementCount);
+                $this->stderr($result);
+            }
+        }
 
-
-        $promise = $reindexEntryRequestPool->promise();
-        $promise->wait();
         Console::endProgress();
 
         return $errorCount === 0 ? ExitCode::OK : ExitCode::UNSPECIFIED_ERROR;
-    }
-
-    protected function printError(ResponseInterface $response, int $requestIndex)
-    {
-        $this->stderr(PHP_EOL . PHP_EOL);
-        $this->stderr(sprintf(
-            '%s%1$s  Request #%d failed: %d %s%1$s',
-            PHP_EOL,
-            $requestIndex,
-            $response->getStatusCode(),
-            $response->getReasonPhrase()
-        ), Console::FG_GREY, Console::BG_RED, Console::BOLD);
-
-        // Try to display details if the response is JSON-encoded
-        $decodedResponse = json_decode($response->getBody()->getContents(), true);
-
-        if (json_last_error() !== JSON_ERROR_NONE) {
-            // Can't do much more!
-            return;
-        }
-
-        if (isset($decodedResponse['error'])) {
-            $this->stderr('  ' . $decodedResponse['error'] . PHP_EOL, Console::BG_RED, Console::FG_GREY);
-        }
-        if (isset($decodedResponse['previousMessage'])) {
-            $this->stderr('  ↳ Caused by: ' . $decodedResponse['previousMessage'] . PHP_EOL, Console::BG_RED, Console::FG_GREY);
-        }
-        $this->stderr(PHP_EOL);
-
-        if (isset($decodedResponse['trace'])) {
-            $this->stderr(PHP_EOL . 'Stack trace:' . PHP_EOL);
-            $this->stderr($decodedResponse['trace'] . PHP_EOL);
-        }
-        if (isset($decodedResponse['previousTrace'])) {
-            $this->stderr(PHP_EOL . 'Previous exception stack trace:' . PHP_EOL);
-            $this->stderr($decodedResponse['previousTrace'] . PHP_EOL);
-        }
-
-        $this->stderr(PHP_EOL);
     }
 }